feat: Optimize LIFF library handling in Nuxt configuration
This commit adds configuration to optimize the inclusion of the @line/liff library in the Nuxt build process. It ensures that the library is only included as a dependency when necessary and prevents its execution during server-side rendering when using the static Nitro preset. These changes enhance the application's performance and maintainability in server-side environments.